pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "Advent of code puzzle repo"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "html-to-markdown>=2.14.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
